.NET 6 upgrade, code cleanup
