+++ SuperCalculator/Views/MainForm.resx
